introduce a new test file for the __spy__ module
